+++ android_client/domain_api/src/main/java/smarthome/client/domain/api/location/usecases/SetupGeofenceUseCase.kt
@@ -1,5 +1,5 @@
 package smarthome.client.domain.api.location.usecases
 
 interface SetupGeofenceUseCase {
-    suspend fun execute(lat: Long, long: Long, radius: Int)
+    suspend fun execute(lat: Double, long: Double, radius: Int)
 }
